+++ system/lib/libc/emscripten_console.c
@@ -41,25 +41,25 @@ void emscripten_console_warnf(const char* fmt, ...) {
   va_end(ap);
 }
 
-void _emscripten_outf(const char* fmt, ...) {
+void emscripten_outf(const char* fmt, ...) {
   va_list ap;
   va_start(ap, fmt);
-  vlogf(fmt, ap, &_emscripten_out);
+  vlogf(fmt, ap, &emscripten_out);
   va_end(ap);
 }
 
-void _emscripten_errf(const char* fmt, ...) {
+void emscripten_errf(const char* fmt, ...) {
   va_list ap;
   va_start(ap, fmt);
-  vlogf(fmt, ap, &_emscripten_err);
+  vlogf(fmt, ap, &emscripten_err);
   va_end(ap);
 }
 
 #ifndef NDEBUG
-void _emscripten_dbgf(const char* fmt, ...) {
+void emscripten_dbgf(const char* fmt, ...) {
   va_list ap;
   va_start(ap, fmt);
-  vlogf(fmt, ap, &_emscripten_dbg);
+  vlogf(fmt, ap, &emscripten_dbg);
   va_end(ap);
 }
 #endif